% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GenerateKeeper.R
\name{convertKeeperToTable}
\alias{convertKeeperToTable}
\title{Covert Keeper profiles to a table}
\usage{
convertKeeperToTable(keeper, removePersonId = FALSE)
}
\arguments{
\item{keeper}{Keeper profiles as created using the \code{\link[=generateKeeper]{generateKeeper()}} function.}

\item{removePersonId}{Remove person IDs from the output? Can be set to \code{TRUE} for privacy reasons.}
}
\value{
A tibble with one row per person, and text columns for each Keeper category. Can be used for manual review.
}
\description{
Covert Keeper profiles to a table
}
