% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createKEEPER.R
\name{createKeeper}
\alias{createKeeper}
\title{Export person level data from OMOP CDM tables for eligible persons in the cohort.}
\usage{
createKeeper(
  connectionDetails = NULL,
  connection = NULL,
  cohortDatabaseSchema = NULL,
  cdmDatabaseSchema,
  tempEmulationSchema = getOption("sqlRenderTempEmulationSchema"),
  cohortTable = "cohort",
  cohortDefinitionId,
  cohortName = NULL,
  sampleSize = 20,
  personIds = NULL,
  databaseId,
  assignNewId = FALSE,
  useAncestor = TRUE,
  doi,
  comorbidities,
  symptoms,
  alternativeDiagnosis,
  drugs,
  diagnosticProcedures,
  measurements,
  treatmentProcedures,
  complications
)
}
\arguments{
\item{connectionDetails}{An R object of type \code{connectionDetails} created using the
\code{\link[DatabaseConnector:createConnectionDetails]{DatabaseConnector::createConnectionDetails()}} function. Not
required of \code{connection} is provided.}

\item{connection}{The connection to the database server created using
\code{\link[DatabaseConnector:connect]{DatabaseConnector::connect()}}. Not required if \code{connectionDetails}
is provided.}

\item{cohortDatabaseSchema}{The name of the database schema that is the location where the
cohort to review is stored.}

\item{cdmDatabaseSchema}{The name of the database schema that contains the OMOP CDM
instance. Requires read permissions to this database. On SQL
Server, this should specify both the database and the schema,
so for example 'cdm_instance.dbo'.}

\item{tempEmulationSchema}{Some database platforms like Oracle and Impala do not truly support temp tables. To
emulate temp tables, provide a schema with write privileges where temp tables
can be created.}

\item{cohortTable}{The tablename that contains the cohort to review.}

\item{cohortDefinitionId}{The cohort id to extract records.}

\item{cohortName}{(optional) Cohort Name}

\item{sampleSize}{(Optional, default = 20) The number of persons to randomly sample. Ignored, if personId is given.}

\item{personIds}{(Optional) A vector of personId's to look for in Cohort table and CDM.}

\item{databaseId}{A short string for identifying the database (e.g. 'Synpuf'). This will be displayed
in shiny app to toggle between databases. Should not have space or underscore (_).}

\item{assignNewId}{(Default = FALSE) Do you want to assign a newId for persons. This will replace the personId in the source with a randomly assigned newId.}

\item{useAncestor}{keeperOutput: a switch for using concept_ancestor to retrieve relevant terms vs using verbatim strings of codes}

\item{doi}{keeperOutput: input vector of concept_ids for disease of interest}

\item{comorbidities}{keeperOutput: input vector of concept_ids for comorbidities associated with the disease of interest (such as smoking or hyperlipidemia for diabetes)}

\item{symptoms}{keeperOutput: input vector of concept_ids for symptoms associated with the disease of interest (such as weight gain or loss for diabetes)}

\item{alternativeDiagnosis}{keeperOutput: input vector of concept_ids for competing diagnosis within a month after the index date}

\item{drugs}{keeperOutput: input vector of concept_ids for drug exposures relevant to the disease of interest, to be used for prior exposures and treatment after the index date.
You may input drugs that are used to treat disease of interest and drugs used to treat alternative diagnosis}

\item{diagnosticProcedures}{keeperOutput: input vector of concept_ids for diagnostic procedures relevant to the condition of interest within a month prior and after the index date}

\item{measurements}{keeperOutput: input vector of concept_ids for lab tests relevant to the disease of interest within a month prior and after the index date}

\item{treatmentProcedures}{keeperOutput: input vector of concept_ids for treatment procedures relevant to the disease of interest within a month after the index date}

\item{complications}{keeperOutput: input vector of concept_ids for complications of the disease of interest within a year after the index date}
}
\value{
Output is a data frame with one row per patient, with the following information per patient:
\itemize{
\item demographics (age, gender);
\item visit_context: information about visits overlapping with the index date (day 0) formatted as the type of visit and its duration;
\item observation_period: information about overlapping OBSERVATION_PERIOD formatted as days prior - days after the index date;
\item presentation: all records in CONDITION_OCCURRENCE on day 0 with corresponding type and status;
\item comorbidities: records in CONDITION_ERA and OBSERVATION that were selected as comorbidities and risk factors within all time prior excluding day 0. The list does not inlcude symptoms, disease of interest and complications;
\item symptoms: records in CONDITION_ERA that were selected as symptoms 30 days prior excluding day 0. The list does not include disease of interest and complications. If you want to see symptoms outside of this window, please place them in complications;
\item prior_disease: records in CONDITION_ERA that were selected as disease of interest or complications all time prior excluding day 0;
\item prior_drugs: records in DRUG_ERA that were selected as drugs of interest all time prior excluding day 0 formatted as day of era start and length of drug era;
\item prior_treatment_procedures: records in PROCEDURE_OCCURRENCE that were selected as treatments of interest within all time prior excluding day 0;
\item diagnostic_procedures: records in PROCEDURE_OCCURRENCE that were selected as diagnostic procedures within all time prior excluding day 0;
\item measurements: records in MEASUREMENT that were selected as measurements (lab tests) of interest within 30 days before and 30 days after day 0 formatted as value and unit (if exists) and assessment compared to the reference range provided in MEASUREMENT table (normal, abnormal high and abnormal low);
\item alternative_diagnosis: records in CONDITION_ERA that were selected as alternative (competing) diagnosis within 90 days before and 90 days after day 0. The list does not include disease of interest;
\item after_disease: same as prior_disease but after day 0;
\item after_drugs: same as prior_drugs but after day 0;
\item after_treatment_procedures: same as prior_treatment_procedures but after day 0;
\item death: death record any time after day 0.
}
}
\description{
Use \code{useAncestor = TRUE} to switch from verbatim string of concept_ids vs ancestors. In latter
case, the app will take you concept_ids and include them along with their descendants.

Use \code{sampleSize} to specify desired number of patients to be selected.

Use \code{assignNewId = TRUE} to replace person_id with a new sequence.

Explanation of categories:
\itemize{
\item instantiated cohort with patients of interest in COHORT table or in another table that has the same fields as COHORT;
\item doi: string for disease of interest (ex.: diabetes type I). Hereon, assume a string of concept_ids;
\item symptoms: symptoms of disease of interest or alternative/competing diagnoses (those that you want to see to be able to distinguish your doi from another close disease, ex.: polyuria, weight gain or loss, vision disturbances);
\item comorbidities: relevant diseases that co-occur with doi or alternative/competing diagnoses (ex.: obesity, metabolic syndrome, pancreatic disorders, pregnancy);
\item drugs: drugs, relevant to the disease of interest or those that can be used to treat alternative/competing diagnoses (ex.: insulin, oral glucose lowering drugs);
\item diagnosticProcedures: relevant diagnostic procedures (ex.: ultrasound of pancreas);
\item measurements: relevant lab tests (ex.: islet cell ab, HbA1C, glucose measurement in blood, insulin ab);
\item alternativeDiagnosis: alternative/competing diagnoses (ex.: diabetes type 2, cystic fibrosis, gestational diabetes, renal failure, pancreonecrosis)
\item treatmentProcedures: relevant treatment procedures (ex.: operative procedures on pancreas);
\item complications: relevant complications (ex.: retinopathy, CKD).
}

*note: if no suitable concept_ids exists for an input string, input c(0)
}
\examples{
\dontrun{
connectionDetails <- createConnectionDetails(
  dbms = 'postgresql',
  server = 'ohdsi.com',
  port = 5432,
  user = 'me',
  password = 'secure'
)

createKeeper(
  connectionDetails = connectionDetails,
  databaseId = "Synpuf",
  cdmDatabaseSchema = "dbo",
  cohortDatabaseSchema = "results",
  cohortTable = "cohort",
  cohortDefinitionId = 1234,
  cohortName = "DM type I",
  sampleSize = 100,
  assignNewId = TRUE,
  useAncestor = TRUE,
  doi = c(201820,442793,443238,4016045,4065354,45757392, 4051114, 433968, 375545, 29555009, 
          4209145, 4034964, 380834, 4299544, 4226354, 4159742, 43530690, 433736, 320128, 4170226,
          40443308, 441267, 4163735, 192963, 85828009),
  symptoms = c(4232487, 4229881),
  comorbidities = c(432867, 436670),
  drugs = c(1730370, 21604490, 21601682, 21601855, 21601462, 21600280, 21602728, 1366773, 
            21602689, 21603923, 21603746),
  diagnosticProcedures = c(40756884, 4143852, 2746768, 2746766),
  measurements	= c(3034962, 3000483, 3034962, 3000483, 3004501, 3033408, 3005131, 3024629, 
                   3031266, 3037110, 3009261, 3022548, 3019210, 3025232, 3033819,
                   3000845, 3002666, 3004077, 3026300, 3014737, 3027198, 3025398, 3010300, 
                   3020399, 3007332, 3025673, 3027457, 3010084, 3004410, 3005673),
  alternativeDiagnosis = c(201820,442793,443238,4016045,4065354,45757392, 4051114, 433968, 
                           375545, 29555009, 4209145, 4034964, 380834, 4299544, 4226354, 4159742,
                           43530690, 433736, 320128, 4170226, 40443308, 441267, 4163735, 192963, 
                           85828009),
  treatmentProcedures = c(0),
  complications =  c(201820,442793,443238,4016045,4065354,45757392, 4051114, 433968, 375545, 
                     29555009, 4209145, 4034964, 380834, 4299544, 4226354, 4159742, 43530690, 
                     433736, 320128, 4170226, 40443308, 441267, 4163735, 192963, 85828009)                             
)
}
}
