% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PromptCraft.R
\name{createPrompt}
\alias{createPrompt}
\title{Create the main prompt based on a Keeper output row.}
\usage{
createPrompt(settings, phenotypeName, keeperTableRow)
}
\arguments{
\item{settings}{A settings object as created using \link{createPromptSettings}.}

\item{phenotypeName}{The name of the disease to use in the prompt.}

\item{keeperTableRow}{A single row from the output of \link{convertKeeperToTable}.}
}
\value{
A character string containing the main prompt.
}
\description{
Create the main prompt based on a Keeper output row.
}
