% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PromptCraft.R
\name{createPromptSettings}
\alias{createPromptSettings}
\title{Create settings for generating prompts}
\usage{
createPromptSettings(maxDays = 20, legacy = FALSE)
}
\arguments{
\item{maxDays}{How many days can a single code have? For example,
if \code{maxDays = 5} and there is a measurement code
that appears on more than 5 days, a random sample
of 5 days will be taken. Set to \code{0} if there is no maximum.}

\item{legacy}{IF TRUE, will use the prompt from the Nature Digital Medicine paper.}
}
\value{
A settings object, to be used in \code{\link[=reviewCases]{reviewCases()}}.
}
\description{
Create settings for generating prompts
}
