% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SensitiveCohort.R
\name{createReferenceCohortTableNames}
\alias{createReferenceCohortTableNames}
\title{Create reference cohort table names}
\usage{
createReferenceCohortTableNames(referenceCohortTable)
}
\arguments{
\item{referenceCohortTable}{The name of the cohort table itself. The metadata table name will be derived from this
by appending '_metadata'.}
}
\value{
A list with \code{referenceCohortTable} and \code{referenceCohortMetadataTable}.
}
\description{
Derives a metadata table name from the reference cohort table name in a systematic way.
}
