% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SensitiveCohort.R
\name{createSensitiveAndSpecificCohorts}
\alias{createSensitiveAndSpecificCohorts}
\title{Create a highly-sensitive cohort for a phenotype}
\usage{
createSensitiveAndSpecificCohorts(
  connectionDetails = NULL,
  connection = NULL,
  cdmDatabaseSchema,
  tempEmulationSchema = getOption("sqlRenderTempEmulationSchema"),
  cohortDatabaseSchema,
  cohortTable,
  createCohortTable = FALSE,
  sensitiveCohortId,
  specificCohortId = NULL,
  keeperConceptSets,
  minDoiPersons = 1000,
  minSpecPersons = 1000
)
}
\arguments{
\item{connectionDetails}{An R object of type \code{connectionDetails} created using the
\code{\link[DatabaseConnector:createConnectionDetails]{DatabaseConnector::createConnectionDetails()}} function. Not
required of \code{connection} is provided.}

\item{connection}{The connection to the database server created using
\code{\link[DatabaseConnector:connect]{DatabaseConnector::connect()}}. Not required if \code{connectionDetails}
is provided.}

\item{cdmDatabaseSchema}{The name of the database schema that contains the OMOP CDM
instance. Requires read permissions to this database. On SQL
Server, this should specify both the database and the schema,
so for example 'cdm_instance.dbo'.}

\item{tempEmulationSchema}{Some database platforms like Oracle and Impala do not truly support temp tables. To
emulate temp tables, provide a schema with write privileges where temp tables
can be created.}

\item{cohortDatabaseSchema}{The name of the database schema that is the location where the
cohort will be stored.}

\item{cohortTable}{The table name where the cohort will be stored.}

\item{createCohortTable}{Create the cohort table. If \code{TRUE} and the table already exists it will first be
deleted.}

\item{sensitiveCohortId}{The cohort definition ID that will be used for the sensitive cohort.}

\item{specificCohortId}{The cohort definition ID that will be used for the specific cohort. If not
provided, no specific cohort will be constructed.}

\item{keeperConceptSets}{A data frame with Keeper concept sets as generated by
\code{\link[=generateKeeperConceptSets]{generateKeeperConceptSets()}}.}

\item{minDoiPersons}{Minimum number of persons with DoI concepts. If the count is lower and the
\code{keeperConceptSets} has a 'hypernym' category that is used instead. Else, an error
is thrown.}

\item{minSpecPersons}{Minimum number of persons in the highly-specific cohort. Constraints on the
specific cohort will be relaxed until this count is achieved.}
}
\value{
This function is primarily executed for instantiating the cohort in the cohort table. It does return a data frame
listing the concepts that were used to create the cohort.
}
\description{
Creates a cohort that is highly sensitive. Nobody outside the cohort is expected to have the disease. As a
consequence, specificity inside the cohort is likely low.

This cohort can be used to determine the sensitivity of a cohort definition.

The cohort is constructed by combining two sets:
\enumerate{
\item Anybody who has one of the condition concepts of the disease of interest, indexed on the first occurrence.
\item Anybody who has 'highly specific concepts' from at least two Keeper categories (symptoms, drugs, etc.), indexed
on the first occurrence.
}

Highly specific concepts are concepts where at least 10\% of those who have that concept also have a condition concept
for the disease of interest.
}
