% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StratifiedSampling.R
\name{evaluateCohortUsingStratifiedSample}
\alias{evaluateCohortUsingStratifiedSample}
\title{Evaluate a cohort when using stratified sampling}
\usage{
evaluateCohortUsingStratifiedSample(
  connectionDetails = NULL,
  connection = NULL,
  cohortDatabaseSchema,
  cohortTable,
  cohortDefinitionId,
  referenceCohortDatabaseSchema,
  referenceCohortTableNames,
  referenceCohortDefinitionId
)
}
\arguments{
\item{connectionDetails}{An R object of type \code{connectionDetails} created using the
\code{\link[DatabaseConnector:createConnectionDetails]{DatabaseConnector::createConnectionDetails()}} function. Not
required of \code{connection} is provided.}

\item{connection}{The connection to the database server created using
\code{\link[DatabaseConnector:connect]{DatabaseConnector::connect()}}. Not required if \code{connectionDetails}
is provided.}

\item{cohortDatabaseSchema}{The name of the database schema containing the cohort
to evaluate.}

\item{cohortTable}{The table name containing the cohort to evaluate.}

\item{cohortDefinitionId}{The cohort definition ID of the cohort to evaluate.}

\item{referenceCohortDatabaseSchema}{The name of the database schema containing the reference
cohort (as uploaded by \code{\link[=uploadReferenceCohortUsingStratifiedSample]{uploadReferenceCohortUsingStratifiedSample()}}).}

\item{referenceCohortTableNames}{The table names where the reference cohort and metadata are stored. Should
be created using [createReferenceCohortTableNamesUsingStratifiedSample())].

[createReferenceCohortTableNamesUsingStratifiedSample())]: R:createReferenceCohortTableNamesUsingStratifiedSample())}

\item{referenceCohortDefinitionId}{The cohort definition ID of the reference cohort.}
}
\value{
A tibble with one row, with columns for PPV, sensitivity, and their confidence intervals.
}
\description{
Computes operating characteristics (sensitivity and positive predictive value) of a cohort definition by comparing it
against a reference cohort created from LLM review of KEEPER profiles. Assumes stratified sampling was used, and uses
the Begg and Greenes framework for optimal power.
}
