% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Shiny.R
\name{launchReviewerApp}
\alias{launchReviewerApp}
\title{Launch the reviewer Shiny app}
\usage{
launchReviewerApp(keeper, decisionsFileName)
}
\arguments{
\item{keeper}{The output of the \code{\link[=generateKeeper]{generateKeeper()}} function.}

\item{decisionsFileName}{The location where the decisions made by the reviewer wil be written.}
}
\value{
Returns nothing. Called for launching the Shiny app.
}
\description{
Launch the reviewer Shiny app
}
