% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ResponseParser.R
\name{parseLlmResponse}
\alias{parseLlmResponse}
\title{Parse the response of a LLM}
\usage{
parseLlmResponse(response, noMatchIsDontKnow = TRUE)
}
\arguments{
\item{response}{The response of a LLM to the system prompt with prompt generated by
\code{createSystemPrompt()} and \code{createPrompt()}, respectively.}

\item{noMatchIsDontKnow}{If the response doesn't fit any predefined pattern, should we return
"I don't know"?}
}
\value{
Returns a character string with one of the following values:
\itemize{
\item "yes": Yes, the patient has the disease.
\item "no": No, the patient does not have the disease.
\item "I don't know": The LLM cannot decide whether the patient has the disease.
\item NA: There was a problem parsing the LLM's response.
}
}
\description{
Parse the response of a LLM
}
\examples{
parseLlmResponse("Summary: yes")
parseLlmResponse("Summary: It is unclear whether the patient has the disease.")

}
