% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SensitiveCohort.R
\name{predictPersonProbabilities}
\alias{predictPersonProbabilities}
\title{Create per-person predicted probabilities}
\usage{
predictPersonProbabilities(
  connectionDetails = NULL,
  connection = NULL,
  cdmDatabaseSchema,
  tempEmulationSchema = getOption("sqlRenderTempEmulationSchema"),
  cohortDatabaseSchema,
  cohortTable,
  sensitiveCohortId,
  specificCohortId,
  maxCohortSizeForFitting = 25000,
  maxCores = 10
)
}
\arguments{
\item{connectionDetails}{An R object of type \code{connectionDetails} created using the
\code{\link[DatabaseConnector:createConnectionDetails]{DatabaseConnector::createConnectionDetails()}} function. Not
required of \code{connection} is provided.}

\item{connection}{The connection to the database server created using
\code{\link[DatabaseConnector:connect]{DatabaseConnector::connect()}}. Not required if \code{connectionDetails}
is provided.}

\item{cdmDatabaseSchema}{The name of the database schema that contains the OMOP CDM
instance. Requires read permissions to this database. On SQL
Server, this should specify both the database and the schema,
so for example 'cdm_instance.dbo'.}

\item{tempEmulationSchema}{Some database platforms like Oracle and Impala do not truly support temp tables. To
emulate temp tables, provide a schema with write privileges where temp tables
can be created.}

\item{cohortDatabaseSchema}{The name of the database schema that is the location where the cohorts are stored.}

\item{cohortTable}{The table name where the cohorts are stored.}

\item{sensitiveCohortId}{The cohort definition ID for the sensitive cohort.}

\item{specificCohortId}{The cohort definition ID for the specific cohort.}

\item{maxCohortSizeForFitting}{The maximum cohort size used for fitting the predictive model. If either the specific
cohort or the negative cohort are larger, they will be downsampled.}

\item{maxCores}{Maximum number of CPU cores used to fit the model.}
}
\value{
Does not return anything. Is executed for the side-effect of populating the \code{probabilities} colums in the cohort
table.
}
\description{
Create per-person predicted probabilities for whether they are a case and stores this in a \code{probability} column in
the cohort table. If the column does not yet exist it will be added.

Takes the following steps:
\enumerate{
\item Samples training population from the specific cohort (positives) and the complement of the sensitive cohort (negatives),
The index date for the negatives is sampled from visit dates such that the distribution of index year and time
since observation start is (roughly) equal.
\item Construct features for the training population, automatically removing features that are highly predictive by
themselves.
\item Fit an L1 logistic regression model.
\item Use the model to predict probabilities for the sensitive cohort.#'
}
}
