% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SensitiveCohort.R
\name{uploadReferenceCohort}
\alias{uploadReferenceCohort}
\title{Upload a reference cohort}
\usage{
uploadReferenceCohort(
  connectionDetails = NULL,
  connection = NULL,
  tempEmulationSchema = getOption("sqlRenderTempEmulationSchema"),
  referenceCohortDatabaseSchema,
  referenceCohortTable,
  referenceCohortDefinitionId,
  createReferenceCohortTable = FALSE,
  reviews
)
}
\arguments{
\item{connectionDetails}{An R object of type \code{connectionDetails} created using the
\code{\link[DatabaseConnector:createConnectionDetails]{DatabaseConnector::createConnectionDetails()}} function. Not
required of \code{connection} is provided.}

\item{connection}{The connection to the database server created using
\code{\link[DatabaseConnector:connect]{DatabaseConnector::connect()}}. Not required if \code{connectionDetails}
is provided.}

\item{tempEmulationSchema}{Some database platforms like Oracle and Impala do not truly support temp tables. To
emulate temp tables, provide a schema with write privileges where temp tables
can be created.}

\item{referenceCohortDatabaseSchema}{The name of the database schema where the reference
cohort will be stored.}

\item{referenceCohortTable}{The table name where the reference cohort will be stored.}

\item{referenceCohortDefinitionId}{The cohort definition ID that will be used for the
reference cohort.}

\item{createReferenceCohortTable}{Create the reference cohort table. If \code{TRUE} and the table
already exists it will first be deleted.}

\item{reviews}{A data frame containing reviews as generated by \code{\link[=reviewCases]{reviewCases()}}.}
}
\value{
This function does not return a value. It is called for the side effect of uploading
the reference cohort to the database.
}
\description{
Uploads a reference cohort based on chart reviews to a database table. The reference cohort
contains the reviewed persons with their case status, and can be used to evaluate the operating
characteristics of a cohort definition using \code{\link[=computeCohortOperatingCharacteristics]{computeCohortOperatingCharacteristics()}}.
}
