% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GenerateKeeper.R
\name{convertKeeperToTable}
\alias{convertKeeperToTable}
\title{Convert Keeper profiles to a table}
\usage{
convertKeeperToTable(keeper, removePii = FALSE)
}
\arguments{
\item{keeper}{Keeper profiles as created using the \code{\link[=generateKeeper]{generateKeeper()}} function.}

\item{removePii}{Remove person identifiable information such as person ID and cohort start date from the output?
patient data later.}
}
\value{
A tibble with one row per person, and text columns for each Keeper category. Can be used for manual review.
}
\description{
Convert Keeper profiles to a table
}
