% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StratifiedSampling.R
\name{createReferenceCohortTableNamesUsingStratifiedSample}
\alias{createReferenceCohortTableNamesUsingStratifiedSample}
\title{Create reference cohort table names when using stratified sampling}
\usage{
createReferenceCohortTableNamesUsingStratifiedSample(referenceCohortTable)
}
\arguments{
\item{referenceCohortTable}{The name of the cohort table itself.}
}
\value{
A list with \code{referenceCohortTable},  \code{referenceCohortMetadataTable}, and \code{referenceSensitiveCohortTable}.
}
\description{
Derives a metadata table name and sensitive cohort table name  from the reference cohort table name in a systematic
way.
}
