% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PromptCraft.R
\name{createSystemPrompt}
\alias{createSystemPrompt}
\title{Create a system prompt for a LLM}
\usage{
createSystemPrompt(settings, diseaseName)
}
\arguments{
\item{settings}{A settings object as created using \code{createPromptSettings()}.}

\item{diseaseName}{The name of the disease to use in the prompt.}
}
\value{
A character string with the system prompt.
}
\description{
Create a system prompt for a LLM
}
