% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GenerateKeeper.R
\name{generateKeeper}
\alias{generateKeeper}
\title{Create a sensitive cohort}
\usage{
generateKeeper(
  connectionDetails = NULL,
  connection = NULL,
  cdmDatabaseSchema,
  tempEmulationSchema = getOption("sqlRenderTempEmulationSchema"),
  cohortDatabaseSchema,
  cohortTable,
  cohortDefinitionId,
  sampleSize = 20,
  personIds = NULL,
  keeperConceptSets,
  useDescendants = TRUE,
  removePersonId = FALSE
)
}
\arguments{
\item{connectionDetails}{An R object of type \code{connectionDetails} created using the
\code{\link[DatabaseConnector:createConnectionDetails]{DatabaseConnector::createConnectionDetails()}} function. Not
required of \code{connection} is provided.}

\item{connection}{The connection to the database server created using
\code{\link[DatabaseConnector:connect]{DatabaseConnector::connect()}}. Not required if \code{connectionDetails}
is provided.}

\item{cdmDatabaseSchema}{The name of the database schema that contains the OMOP CDM
instance. Requires read permissions to this database. On SQL
Server, this should specify both the database and the schema,
so for example 'cdm_instance.dbo'.}

\item{tempEmulationSchema}{Some database platforms like Oracle and Impala do not truly support temp tables. To
emulate temp tables, provide a schema with write privileges where temp tables
can be created.}

\item{cohortDatabaseSchema}{The name of the database schema that is the location where the
cohort to review is stored.}

\item{cohortTable}{The tablename that contains the cohort to review.}

\item{cohortDefinitionId}{The cohort definition ID that will be used for the cohort.}

\item{keeperConceptSets}{A data frame with Keeper concept sets as generated by
\code{\link[=generateKeeperConceptSets]{generateKeeperConceptSets()}}.}

\item{useDescendants}{Include the  descendants of the concepts specified in the \code{keeperConceptSets}, or
use only the verbatim concepts?}

\item{removePersonId}{Remove person IDs from the output? Can be set to \code{TRUE} for privacy reasons.}

\item{createCohortTable}{Create the cohort table. If \code{TRUE} and the table already exists it will first be
deleted.}
}
\value{
Returns a list of tibbles with data in the various KEEPER categories.
}
\description{
Creates a cohort that is highly sensitive. Nobody outside the cohort is expected to have the disease. As a
consequence, specificity inside the cohort is likely low.

This cohort can be used to determine the sensitivity of a cohort definition.
}
