% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GenerateConceptSets.R
\name{generateKeeperConceptSets}
\alias{generateKeeperConceptSets}
\title{Generate Keeper input concept sets}
\usage{
generateKeeperConceptSets(
  phenotype,
  phenotypeDefinition = NULL,
  client,
  vocabConnectionDetails,
  vocabDatabaseSchema
)
}
\arguments{
\item{phenotype}{A text string denoting the condition of interest, e.g. 'Type I Diabetes Mellitus
(T1DM)'. This string is used as input for the LLM.}

\item{phenotypeDefinition}{Optionally prove a text blob with the definition and any other information about the
phenotype.}

\item{client}{An LLM client created using the \code{ellmer} package.}

\item{vocabConnectionDetails}{Connection details for a database server hosting the OHDSI Vocabulary. Should be
created using the \link[DatabaseConnector]{createConnectionDetails} function in the
\code{DatabaseConnector} package.}

\item{vocabDatabaseSchema}{The name of the database schema on the server where the vocabulary tables are located.}
}
\value{
A data frame with the following columns:
\itemize{
\item \code{conceptId}
\item \code{conceptName}
\item \code{vocabularyId}
\item \code{conceptSetName}: 'doi', 'symptoms', ...
\item \code{target}: 'Disease of interest' or 'Alternative diagnoses'.
}
}
\description{
Generate the concept sets used as input for Keeper. This function uses LLMs, the OHDSI vocabulary vector store, and
Phoebe to populate concept sets for the disease of interest, symptoms, treatments etc. of the given medical
condition.
}
