% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Keeper.R
\docType{package}
\name{Keeper-package}
\alias{Keeper}
\alias{Keeper-package}
\title{Keeper: An R package to review patient profiles for phenotype validation}
\description{
An R package to review patient profiles for phenotype validation.
}
\author{
\strong{Maintainer}: Ostropolets Anna \email{ostropolets@ohdsi.org}

Authors:
\itemize{
  \item Schuemie Martjn
}

Other contributors:
\itemize{
  \item Observational Health Data Science and Informatics [copyright holder]
}

}
\keyword{internal}
