% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Shiny.R
\name{launchReviewerApp}
\alias{launchReviewerApp}
\title{Launch the reviewer Shiny app}
\usage{
launchReviewerApp(keeper, keeperConceptSets, decisionsFileName)
}
\arguments{
\item{keeper}{The output of the \code{\link[=generateKeeper]{generateKeeper()}} function.}

\item{keeperConceptSets}{The output of the \code{\link[=generateKeeperConceptSets]{generateKeeperConceptSets()}} function.}

\item{decisionsFileName}{The location of the CSV file where the decisions made by the reviewer will be written. If
the file doesn't exist it will be created. If it does exist, it will be loaded and can be
edited in the Shiny app.}
}
\value{
Returns nothing. Called for launching the Shiny app.
}
\description{
Launch the reviewer Shiny app
}
