% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LlmLoop.R
\name{reviewCases}
\alias{reviewCases}
\title{Review Keeper profiles using an LLM}
\usage{
reviewCases(
  keeper,
  settings = createPromptSettings(),
  phenotypeName = NULL,
  client,
  cacheFolder
)
}
\arguments{
\item{keeper}{Output from the \code{\link[=generateKeeper]{generateKeeper()}} function.}

\item{settings}{Prompt creating settings as created using the \link{createPromptSettings} function.}

\item{phenotypeName}{The name of the disease to use in the prompt. If not provided, the name in the Keeper input will
be used.}

\item{client}{An LLM client created using the \code{ellmer} package.}

\item{cacheFolder}{A folder where the LLM responses are cached. If the process terminates for some
reason, it can pick up where it left off using the cache.}
}
\value{
A tibble with these columns:
\itemize{
\item \code{generatedId}
\item \code{isCase}, with possible values "yes" or "no",
\item \code{certainty}, certainty of the LLM in its decision, can be "high" or "low".
\item \code{justification}, written by the LLM.
\item \code{cohortPrevalence}, prevalence of the cohort in the entire population.
\item \code{model}, the LMM used to review.
\item \code{keeperVersion}, the version of the Keeper package.
}

When the Keeper profiles were generated with \code{removePii = FALSE}, the following columns are also included:
\itemize{
\item \code{personId}
\item \code{cohortStartDate}
}
}
\description{
Review Keeper profiles using an LLM
}
