% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LlmLoop.R
\name{reviewCases}
\alias{reviewCases}
\title{Review cases using an LLM}
\usage{
reviewCases(
  keeper,
  settings = createPromptSettings(),
  diseaseName,
  client,
  cacheFolder = NULL
)
}
\arguments{
\item{keeper}{Output from the \link{createKeeper} function.}

\item{settings}{Prompt creating settings as created using the \link{createPromptSettings} function.}

\item{diseaseName}{The name of the disease to use in the prompt.}

\item{client}{An LLM client created using the \code{ellmer} package.}

\item{cacheFolder}{(Optional) A folder where the LLM responses are cached. If the process terminates for some
reason, it can pick up where it left off using the cache.}
}
\value{
A data frame with two columns:
\enumerate{
\item \code{personId}
\item \code{isCase}, with possible values "yes", "no", or "I don't know".
}
}
\description{
Review cases using an LLM
}
