% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StratifiedSampling.R
\name{reviewCasesUsingStratifiedSampling}
\alias{reviewCasesUsingStratifiedSampling}
\title{Review Keeper profiles using an LLM and stratified sampling}
\usage{
reviewCasesUsingStratifiedSampling(
  connectionDetails = NULL,
  connection = NULL,
  cdmDatabaseSchema,
  tempEmulationSchema = getOption("sqlRenderTempEmulationSchema"),
  sensitiveCohortDatabaseSchema = NULL,
  sensitiveCohortTable,
  sensitiveCohortDefinitionId,
  phase1SampleSize = 100,
  targetRelativeError = 0.1,
  targetTrueCases = 150,
  keeperConceptSets,
  useDescendants = TRUE,
  settings = createPromptSettings(),
  phenotypeName = NULL,
  clinicalDefinition = NULL,
  client,
  cacheFolder
)
}
\arguments{
\item{connectionDetails}{An R object of type \code{connectionDetails} created using the
\code{\link[DatabaseConnector:createConnectionDetails]{DatabaseConnector::createConnectionDetails()}} function. Not
required of \code{connection} is provided.}

\item{connection}{The connection to the database server created using
\code{\link[DatabaseConnector:connect]{DatabaseConnector::connect()}}. Not required if \code{connectionDetails}
is provided.}

\item{cdmDatabaseSchema}{The name of the database schema that contains the OMOP CDM
instance. Requires read permissions to this database. On SQL
Server, this should specify both the database and the schema,
so for example 'cdm_instance.dbo'.}

\item{tempEmulationSchema}{Some database platforms like Oracle and Impala do not truly support temp tables. To
emulate temp tables, provide a schema with write privileges where temp tables
can be created.}

\item{sensitiveCohortDatabaseSchema}{Name of the database schema where the sensitive cohort is stored.}

\item{sensitiveCohortTable}{Name of the table where the sensitive cohort is stored.}

\item{sensitiveCohortDefinitionId}{Cohort definition ID of the sensitive cohort.}

\item{phase1SampleSize}{The size of the sample in each stratum in phase 1.}

\item{targetRelativeError}{The target relative error in our metrics when computing the required sample size.}

\item{targetTrueCases}{The target number of true cases in the final sample, when computing the required sample size.}

\item{keeperConceptSets}{A data frame with Keeper concept sets as generated by
\code{\link[=generateKeeperConceptSets]{generateKeeperConceptSets()}}.}

\item{useDescendants}{Include the  descendants of the concepts specified in the \code{keeperConceptSets}, or
use only the verbatim concepts?}

\item{settings}{Prompt creating settings as created using the \link{createPromptSettings} function.}

\item{phenotypeName}{The name of the disease to use in the prompt. If not provided, the name in the Keeper input will
be used.}

\item{clinicalDefinition}{Optionally prove a text blob with the definition and any other information about the
phenotype.}

\item{client}{An LLM client created using the \code{ellmer} package.}

\item{cacheFolder}{A folder where the LLM responses are cached. If the process terminates for some
reason, it can pick up where it left off using the cache.}
}
\value{
An object (list) with two components:
\itemize{
\item \code{llmReviews}: The same output as \code{generateKeeper()}, for the sampled persons.
\item \code{stratification}: A mapping from the stratification information columns to the strata IDs.
}

This object is intended to be used in \code{uploadReferenceCohortUsingStratifiedSample()}.
}
\description{
Uses an LLM to adjudicate cases, while relying on stratified sampling. This process follows these steps:
\enumerate{
\item Create strata using the stratification stored in the sensitive cohort. Smaller strata are merged.
\item Phase 1: Take a small sample per stratum, and adjudicate using the LLM.
\item Phase 2: Based on the phase 1 adjudications, estimate prevalence per stratum, and use Neyman allocation to achieve
the specified target relative error and number of true cases. Adjudicate using the LLM.
The output combined the adjudications from phase 1 and 2.
}
}
