% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SensitiveCohort.R
\name{uploadReferenceCohort}
\alias{uploadReferenceCohort}
\title{Upload a reference cohort}
\usage{
uploadReferenceCohort(
  connectionDetails = NULL,
  connection = NULL,
  tempEmulationSchema = getOption("sqlRenderTempEmulationSchema"),
  referenceCohortDatabaseSchema,
  referenceCohortTableNames,
  referenceCohortDefinitionId,
  createReferenceCohortTables = FALSE,
  reviews
)
}
\arguments{
\item{connectionDetails}{An R object of type \code{connectionDetails} created using the
\code{\link[DatabaseConnector:createConnectionDetails]{DatabaseConnector::createConnectionDetails()}} function. Not
required of \code{connection} is provided.}

\item{connection}{The connection to the database server created using
\code{\link[DatabaseConnector:connect]{DatabaseConnector::connect()}}. Not required if \code{connectionDetails}
is provided.}

\item{tempEmulationSchema}{Some database platforms like Oracle and Impala do not truly support temp tables. To
emulate temp tables, provide a schema with write privileges where temp tables
can be created.}

\item{referenceCohortDatabaseSchema}{The name of the database schema where the reference
cohort will be stored.}

\item{referenceCohortTableNames}{The table names where the reference cohort and metadata will be stored. Should
be created using \code{\link[=createReferenceCohortTableNames]{createReferenceCohortTableNames()}}.}

\item{referenceCohortDefinitionId}{The cohort definition ID that will be used for the
reference cohort.}

\item{createReferenceCohortTables}{Create the reference cohort and metadata tables? If \code{TRUE} and the tables
already exists they will first be deleted.}

\item{reviews}{A data frame containing reviews as generated by \code{\link[=reviewCases]{reviewCases()}}.}
}
\value{
This function does not return a value. It is called for the side effect of uploading
the reference cohort to the database.
}
\description{
A reference cohort is typically a large sample (e.g. 10,000 persons) of a highly-sensitive cohort (as created using
\code{\link[=createSensitiveCohort]{createSensitiveCohort()}}), reviewed by an LLM (using \code{\link[=reviewCases]{reviewCases()}}).

The reference cohort can be used to compute operating characteristics of a cohort definition for the same phenotype
using \code{\link[=computeCohortOperatingCharacteristics]{computeCohortOperatingCharacteristics()}}.
}
