% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StratifiedSampling.R
\name{uploadReferenceCohortUsingStratifiedSample}
\alias{uploadReferenceCohortUsingStratifiedSample}
\title{Upload a reference cohort when using stratified sampling}
\usage{
uploadReferenceCohortUsingStratifiedSample(
  connectionDetails = NULL,
  connection = NULL,
  tempEmulationSchema = getOption("sqlRenderTempEmulationSchema"),
  sensitiveCohortDatabaseSchema,
  sensitiveCohortTable,
  sensitiveCohortDefinitionId,
  referenceCohortDatabaseSchema,
  referenceCohortTableNames,
  referenceCohortDefinitionId,
  createReferenceCohortTables = FALSE,
  stratifiedReviews
)
}
\arguments{
\item{connectionDetails}{An R object of type \code{connectionDetails} created using the
\code{\link[DatabaseConnector:createConnectionDetails]{DatabaseConnector::createConnectionDetails()}} function. Not
required of \code{connection} is provided.}

\item{connection}{The connection to the database server created using
\code{\link[DatabaseConnector:connect]{DatabaseConnector::connect()}}. Not required if \code{connectionDetails}
is provided.}

\item{tempEmulationSchema}{Some database platforms like Oracle and Impala do not truly support temp tables. To
emulate temp tables, provide a schema with write privileges where temp tables
can be created.}

\item{sensitiveCohortDatabaseSchema}{Name of the database schema where the sensitive cohort is stored.}

\item{sensitiveCohortTable}{Name of the table where the sensitive cohort is stored.}

\item{sensitiveCohortDefinitionId}{Cohort definition ID of the sensitive cohort.}

\item{referenceCohortDatabaseSchema}{The name of the database schema where the reference
cohort will be stored.}

\item{referenceCohortTableNames}{The table names where the reference cohort, metadata, and stratified sensitve
cohort will be stored. Should be created using
\code{\link[=createReferenceCohortTableNamesUsingStratifiedSample]{createReferenceCohortTableNamesUsingStratifiedSample()}}.}

\item{referenceCohortDefinitionId}{The cohort definition ID that will be used for the
reference cohort.}

\item{createReferenceCohortTables}{Create the reference cohort and metadata tables? If \code{TRUE} and the tables
already exists they will first be deleted.}

\item{stratifiedReviews}{An object as generated by \code{\link[=reviewCasesUsingStratifiedSampling]{reviewCasesUsingStratifiedSampling()}}.}
}
\value{
This function does not return a value. It is called for the side effect of uploading
the reference cohort to the database.
}
\description{
A reference cohort is a large sample of a highly-sensitive cohort reviewed by an LLM. Here we assume stratified
sampling was performed using \code{\link[=reviewCasesUsingStratifiedSampling]{reviewCasesUsingStratifiedSampling()}}.

The reference cohort can be used to compute operating characteristics of a cohort definition for the same phenotype
using \code{\link[=evaluateCohortUsingStratifiedSample]{evaluateCohortUsingStratifiedSample()}}.
}
